+++ CHANGELOG.md
@@ -3,3 +3,12 @@
 FEATURES:
 
 - Added "scope" parameter for the rewrite option
+
+## 1.3.0
+
+FEATURES:
+
+- Added "scope" parameter for the rewrite option
+- Added "forward_host_header" parameter for origin option 
+- Added "browser" parameter for cache -> valid option 
+- Added "order" parameter only for location option